+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### [Mailer]
 
 - Bugfixing possible error on newsletter list editing if custom sorting was used. remp/remp#516
+- Fix broken autologin parameter in unsubscribe email links. remp/remp#518
 
 ---